Separate out card and context meny handling from selectable_list_view
